+++ R/series__series.R
@@ -50,75 +50,6 @@ wrap_s = function(x) {
   if (inherits(x, "Series")) x else pl$Series(x)
 }
 
-
-
-#' Print Series
-#' @export
-#' @param x Series
-#' @param ... not used
-#' @keywords internal
-#' @name Series_print
-#'
-#' @return invisible(self)
-#' @examples print(pl$Series(1:3))
-print.Series = function(x, ...) {
-  cat("polars Series: ")
-  x$print()
-  invisible(x)
-}
-
-#' Print Series
-#' @rdname Series_print
-#' @return self
-#'
-#' @examples pl$Series(1:3)
-Series_print = function() {
-  .pr$Series$print(self)
-  invisible(self)
-}
-
-#' @title auto complete $-access into a polars object
-#' @description called by the interactive R session internally
-#' @param x Series
-#' @param pattern code-stump as string to auto-complete
-#' @return char vec
-#' @export
-#' @inherit .DollarNames.DataFrame return
-#' @keywords internal
-.DollarNames.Series = function(x, pattern = "") {
-  get_method_usages(Series, pattern = pattern)
-}
-
-#' Immutable combine series
-#' @param x a Series
-#' @param ... Series(s) or any object into Series meaning `pl$Series(object)` returns a series
-#' @return a combined Series
-#' @details append datatypes has to match. Combine does not rechunk.
-#' Read more about R vectors, Series and chunks in \code{\link[polars]{docs_translations}}:
-#' @examples
-#' s = c(pl$Series(1:5), 3:1, NA_integer_)
-#' s$chunk_lengths() # the series contain three unmerged chunks
-#' @export
-c.Series = \(x, ...) {
-  l = list2(...)
-  x = x$clone() # clone to retain an immutable api, append_mut is not immutable
-  for (i in seq_along(l)) { # append each element of i being either Series or Into<Series>
-    unwrap(.pr$Series$append_mut(x, wrap_s(l[[i]])), "in $c:")
-  }
-  x
-}
-
-#' Length of series
-#' @param x a Series
-#' @return the length as a double
-#' @export
-length.Series = \(x) x$len()
-
-
-
-
-
-
 #' Create new Series
 #' @description found in api as pl$Series named Series_constructor internally
 #'
@@ -137,12 +68,15 @@ pl$Series = function(x, name = NULL) {
     unwrap("in pl$Series()")
 }
 
-
-
-
-
-
-
+#' Print Series
+#' @rdname Series_print
+#' @return self
+#'
+#' @examples pl$Series(1:3)
+Series_print = function() {
+  .pr$Series$print(self)
+  invisible(self)
+}
 
 #' add Series
 #' @name Series_add
@@ -759,7 +693,7 @@ Series_dtype = method_as_property(function() {
 #' @keywords Series
 #' @return DataType
 #' @aliases Series_flags
-#' @name Series_dtype
+#' @name Series_flags
 #' @details property sorted flags are not settable, use set_sorted
 #' @examples
 #' pl$Series(1:4)$sort()$flags