+++ NEWS.md
@@ -11,6 +11,7 @@
 
 ## What's changed
 
+- `pl$concat()` now also supports `Series`, `Expr` and `LazyFrame` (#407).
 - New method `$unnest()` for `LazyFrame` (#397).
 - New method `$sample()` for `DataFrame` (#399).
 - New method `$meta$tree_format()` to display an `Expr` as a tree (#401).
@@ -20,6 +21,7 @@
   complete in one edge case (#409).
 - New method `$cat$get_categories()` to list unique values of categorical 
   variables (#412).
+- New methods `$fold()` and `$reduce()` to apply an R function rowwise (#403).
 
 # polars 0.8.1